data/requirements.txt
# Core Data Science Libraries
numpy>=1.24.0
pandas>=2.0.0
scipy>=1.11.0

# Machine Learning
scikit-learn>=1.3.0

# Natural Language Processing
nltk>=3.8

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0
wordcloud>=1.9.0
plotly>=5.17.0

# Web Application
streamlit>=1.28.0

# Utility
joblib>=1.3.0

# Advanced Features
textblob>=0.17.0
python-docx>=0.8.11
